pyproject.toml
[project]
name = "civiceye"
version = "0.1.0"
description = "Streamlit app that finds and ranks civic number matches in OpenStreetMap."
readme = "README.md"
requires-python = ">=3.11"
authors = [{ name = "CivicEye Contributors" }]
dependencies = [
    "requests>=2.32.0",
    "streamlit>=1.29.0",
    "pillow>=11.0.0",
]

[project.optional-dependencies]
viz = [
    "sentence-transformers>=3.0.0",
    "torch>=2.1.0",
    "transformers>=4.37.0",
]
docs = [
    "mkdocs>=1.6.0",
    "mkdocs-material>=9.5.0",
]

[project.urls]
Homepage = "https://github.com/USERNAME/REPOSITORY"
Documentation = "https://USERNAME.github.io/REPOSITORY/"
Repository = "https://github.com/USERNAME/REPOSITORY"
Issues = "https://github.com/USERNAME/REPOSITORY/issues"

[build-system]
requires = ["setuptools>=67", "wheel"]
build-backend = "setuptools.build_meta"
